refactor: Missions 스타일 코드 분리
