+++ src/App.tsx
@@ -13,7 +13,7 @@ import DiariesNewMood from '@/pages/Diaries/New/Mood';
 import Errors from '@/pages/Errors';
 import Home from '@/pages/Home';
 import Login from '@/pages/Login';
-import Missions from '@/pages/Missions';
+import Missions from '@/pages/Missions/Missions';
 import OnboardingLayout from '@/pages/Onboarding/Layout';
 import OnboardingStep from '@/pages/Onboarding/Step';
 import AppLayout from './Layout';
